Issue/9 studip auth routes

+++ lib/Routes/UserItems/X5UserItemDelete.php
@@ -12,15 +12,16 @@ class X5UserItemDelete extends JsonApiController
 {
     public function __invoke(Request $request, Response $response, $args)
     {
+        global $perm;
+
+        if (!$perm->have_perm('autor')) {
+            throw new AuthorizationFailedException();
+        };
+
         $user_id = $this->getUser($request)->id;
         if (!$userItem = X5UserItem::findOneBySql('user_id = ? AND item_id = ?', [$user_id, $args['id']])) {
             throw new RecordNotFoundException();
         }
-
-        // TODO: Authorization
-        // if (1 == 2) {
-        //     throw new AuthorizationFailedException();
-        // }
 
         if (!$userItem->delete()) {
             throw new InternalServerError('Could not delete x5list.');
